feat: 더미데이터 별도 파일 생성 및 페이지에 따른 userId전달 수정(근태상세는 라우터 /:id 로 전달
