+++ frontend/src/pages/attendance/AttendanceDetailPage.tsx
@@ -5,7 +5,8 @@ import { useParams } from "react-router-dom";
 
 
 const AttendanceDetailPage = () => {
-   const { id } = useParams();
+  const { id } = useParams();
+  const userId = id ? parseInt(id, 10) : null;
     
 
   return (
@@ -20,7 +21,7 @@ const AttendanceDetailPage = () => {
         py: 4,
         boxSizing: 'border-box', // 패딩 포함 너비 계산
       }}
-    >
+    >1
       {/* 실제 콘텐츠 영역 - 가운데 정렬됨 */}
       <Box
         sx={{
